EBP-18: Implemented refactor of cache functionality as requested in PR review. Removed dead code/comments from tests.
